[gh-41]: move reducers and store into core folder
