--- src/app/reducers/index.js
@@ -1,12 +1,0 @@
-import { combineReducers } from 'redux-immutable'
-import app from './app'
-import searchSkipass from './search-skipass'
-import skipasses from './skipasses'
-
-const rootReducer = combineReducers({
-  app,
-  searchSkipass,
-  skipasses
-})
-
-export default rootReducer
