+++ src/client.js
@@ -7,7 +7,7 @@ import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router'
 import isEmpty from 'lodash.isempty'
 import PouchDB from './app/data-access/PouchDB-client'
-import configureStore from './app/store/configureStore'
+import configureStore from './app/core/store/configureStore'
 import App from './app/App'
 import * as config from './config'
